+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/service/plugin/EntryFilter2Test.java
@@ -22,11 +22,10 @@ package org.apache.pulsar.broker.service.plugin;
 import java.util.List;
 import org.apache.bookkeeper.mledger.Entry;
 import org.apache.commons.collections4.MapUtils;
-import org.apache.pulsar.broker.service.EntryFilter;
 import org.apache.pulsar.broker.service.persistent.PersistentSubscription;
 import org.apache.pulsar.common.api.proto.KeyValue;
 
-public class EntryFilterForTest2 implements EntryFilter {
+public class EntryFilter2Test implements EntryFilter {
     @Override
     public FilterResult filterEntry(Entry entry, FilterContext context) {
         if (context.getMsgMetadata() == null || context.getMsgMetadata().getPropertiesCount() <= 0) {
